requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
openpyxl==3.1.2
google-auth>=2.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
